Refactor image parsing, add comments and update tests
