+++ lib/creek.rb
@@ -3,6 +3,7 @@ require 'creek/book'
 require 'creek/styles/constants'
 require 'creek/styles/style_types'
 require 'creek/styles/converter'
+require 'creek/utils'
 require 'creek/styles'
 require 'creek/drawing'
 require 'creek/sheet'
