+++ lib/creek/drawing.rb
@@ -2,22 +2,32 @@ require 'pathname'
 
 module Creek
   class Drawing
+    include Creek::Utils
 
     COLUMNS = ('A'..'AZ').to_a
 
-    def initialize(book, sheetfile)
+    def initialize(book, drawing_filepath)
       @book = book
-      @sheetfile = sheetfile
+      @drawing_filepath = drawing_filepath
+      @drawings = []
+      @drawings_rels = []
       @images_pathnames = Hash.new { |hash, key| hash[key] = [] }
 
-      load_drawings_and_rels
-      load_images_pathnames_by_cells if has_images?
+      if file_exist?(@drawing_filepath)
+        load_drawings_and_rels
+        load_images_pathnames_by_cells if has_images?
+      end
     end
 
+    ##
+    # Returns false if there are no images in the drawing file or the drawing file does not exist, true otherwise.
     def has_images?
-      @has_images ||= !@drawings.nil? && @drawings.size > 0
+      @has_images ||= !@drawings.empty?
     end
 
+    ##
+    # Extracts images from excel to tmpdir for a cell, if the images are not already extracted (multiple calls or same image file in multiple cells).
+    # Returns array of images as Pathname objects or nil.
     def images_at(cell_name)
       coordinate = calc_coordinate(cell_name)
       pathnames_at_coordinate = @images_pathnames[coordinate]
@@ -36,57 +46,35 @@ module Creek
 
     private
 
+    ##
+    # Transforms cell name to [row, col], e.g. A1 => [0, 0], B3 => [1, 2]
+    # Rows and cols start with 0.
     def calc_coordinate(cell_name)
       col = COLUMNS.index(cell_name.slice /[A-Z]+/)
       row = (cell_name.slice /\d+/).to_i - 1 # rows in drawings start with 0
       [row, col]
     end
 
-    # for saving extracted images from excel
+    ##
+    # Creates/loads temporary directory for extracting images from excel
     def tmpdir
       @tmpdir ||= ::Dir.mktmpdir('creek__drawing')
     end
 
+    ##
+    # Parses drawing and drawing's relationships xmls.
+    # Drawing xml contains relationships ID's and coordinates (row, col).
+    # Drawing relationships xml contains images' locations.
     def load_drawings_and_rels
-      drawing_filepath = extract_drawing_filepath
-      return if drawing_filepath.nil?
-
-      drawing_filepath.sub!('..', 'xl')
-      @drawings = load_drawings(drawing_filepath)
-      @drawings_rels = load_drawings_rels(drawing_filepath)
+      @drawings = parse_xml(@drawing_filepath).css('xdr|twoCellAnchor')
+      drawing_rels_filepath = expand_to_rels_path(@drawing_filepath)
+      @drawings_rels = parse_xml(drawing_rels_filepath).css('Relationships')
     end
 
-    def load_drawings(drawing_filepath)
-      parse_xml(drawing_filepath).css('xdr|twoCellAnchor')
-    end
-
-    def load_drawings_rels(drawing_filepath)
-      drawing_rels_filepath = expand_to_rels_path(drawing_filepath)
-      parse_xml(drawing_rels_filepath).css('Relationships')
-    end
-
-    def extract_drawing_filepath
-      # read drawing relationship ID from the sheet
-      sheet_filepath = "xl/#{@sheetfile}"
-      drawing = parse_xml(sheet_filepath).css('drawing').first
-      return if drawing.nil?
-
-      drawing_rid = drawing.attributes['id'].value
-
-      # read sheet rels to find drawing file location
-      sheet_rels_filepath = expand_to_rels_path(sheet_filepath)
-      parse_xml(sheet_rels_filepath).css("Relationship[@Id='#{drawing_rid}']").first.attributes['Target'].value
-    end
-
-    def expand_to_rels_path(filepath)
-      filepath.sub(/(\/[^\/]+$)/, '/_rels\1.rels')
-    end
-
-    def parse_xml(xml_path)
-      doc = @book.files.file.open(xml_path)
-      Nokogiri::XML::Document.parse(doc)
-    end
-
+    ##
+    # Iterates through the drawings and saves images' paths as Pathname objects to a hash with [row, col] keys.
+    # As multiple images can be located in a single cell, hash values are array of Pathname objects.
+    # One image can be spread across multiple cells (defined with from-row/to-row/from-col/to-col attributes) - same Pathname object is associated to each row-col combination for the range.
     def load_images_pathnames_by_cells
       image_selector = 'xdr:pic/xdr:blipFill/a:blip'.freeze
       row_from_selector = 'xdr:from/xdr:row'.freeze
