+++ spec/drawing_spec.rb
@@ -3,9 +3,9 @@ require './spec/spec_helper'
 describe 'drawing' do
   let(:book) { Creek::Book.new('spec/fixtures/sample-with-images.xlsx') }
   let(:book_no_images) { Creek::Book.new('spec/fixtures/sample.xlsx') }
-  let(:sheetfile) { 'worksheets/sheet1.xml' }
-  let(:drawing) { Creek::Drawing.new(book, sheetfile) }
-  let(:drawing_without_images) { Creek::Drawing.new(book_no_images, sheetfile) }
+  let(:drawingfile) { 'xl/drawings/drawing1.xml' }
+  let(:drawing) { Creek::Drawing.new(book, drawingfile) }
+  let(:drawing_without_images) { Creek::Drawing.new(book_no_images, drawingfile) }
 
   describe '#has_images?' do
     it 'has' do
